Separate BookmarkItem from Bookmarks
https://youtrack.weseek.co.jp/issue/GW-7846
- Create new file for BookmarkItem in Bookmarks folder
- Rename BookmarksItem to BookmarkItem
- Implement BookmarkItem in Bookmarks component
